+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.16.0 - Unreleased
 --------------------------
 - Enh: Added support for hidden files in stream
+- Enh #173: Soft deletion of nested content and restore parent folders on restore a child file/folder
 
 0.15.1 - February 14, 2023
 --------------------------